+++ src/assets/help/changelog.md
@@ -2,6 +2,7 @@
 
 - Fixed an issue where production building recipe selection did not take into account the buildings efficiency on revenue / day and roi metrics (by `lilbit-prun`)
 - Add "Buy From CX" option for XIT Actions for FIO Burn (by `lilbit-prun`)
+- Implement Resource ROI Overview [#130](https://github.com/PRUNplanner/frontend/issues/130)
 
 # 2025-08-12